Add kernel parameters to BLS snippets if kernelopts var isn't present
Since Fedora 33, the kernel command line parameters are not stored in the
GRUB environment variables block and instead are stored in the BLS files.

Update the script to take this into account, otherwise attempting to add a
paramter will not work. For example, the --addconsole option is currently
not adding the expected "console=ttyS0,115200" param to get serial output.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>