Don't add a console param for target none
Currently, if the --addconsole option and used along with --target=none,
the default "ttyS0,115200" console kernel command line paramter is added.

  = No console listed for none, adding default ttyS0,115200

But if the user choses "none" as the target, the image shouldn't make any
assumptions about the platform. Using "ttyS0,115200" as a heuristic if the
console is not set for a particular platform makes sense. But no for none,
in that case it should be provided either by the platform or by the user.

This may even be counterproductive. For example boards that already define
a "stdout-path" property in their "chosen" node, don't even need a console
parameter and adding one will have higher precedence and won't be used the
one that was defined in the Device Tree.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>