Remove count option in dd commands for OMAP boards
The dd commands for the OMAP boards use a "count" option, which limits how
much of the u-boot binaries are written to the storage media. But that is
not really needed since dd will write the size of the bootloader binaries.

What's important is to have the "bs" and "seek" options, because there are
specific offsets on which the MLO and u-boot.img binaries must te placed.

But the "count" may even be harmfull, since for example will cause to not
write the whole u-boot.img binary if its size is bigger than 2 * 384k.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>